LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/forward-already-detached-expected.txt



FAIL forward() in a detached window assert_equals: expected 1 but got 0
